+++ apps/backoffice-v2/src/pages/Entity/components/CallToAction/hooks/useCallToActionLogic/useCallToActionLogic.tsx
@@ -3,9 +3,11 @@ import { useFilterId } from '../../../../../../common/hooks/useFilterId/useFilte
 import { useWorkflowQuery } from '../../../../../../domains/workflows/hooks/queries/useWorkflowQuery/useWorkflowQuery';
 import { useAuthenticatedUserQuery } from '../../../../../../domains/auth/hooks/queries/useAuthenticatedUserQuery/useAuthenticatedUserQuery';
 import { useCaseState } from '../../../Case/hooks/useCaseState/useCaseState';
-import { useUpdateWorkflowByIdMutation } from '../../../../../../domains/workflows/hooks/mutations/useUpdateWorkflowByIdMutation/useUpdateWorkflowByIdMutation';
 import toast from 'react-hot-toast';
 import { useCallback, useState } from 'react';
+import { useApproveTaskByIdMutation } from '../../../../../../domains/entities/hooks/mutations/useApproveTaskByIdMutation/useApproveTaskByIdMutation';
+import { useRejectTaskByIdMutation } from '../../../../../../domains/entities/hooks/mutations/useRejectTaskByIdMutation/useRejectTaskByIdMutation';
+import { useRevisionTaskByIdMutation } from '../../../../../../domains/entities/hooks/mutations/useRevisionTaskByIdMutation/useRevisionTaskByIdMutation';
 
 export const useCallToActionLogic = () => {
   const { entityId } = useParams();
@@ -13,81 +15,12 @@ export const useCallToActionLogic = () => {
   const { data: workflow } = useWorkflowQuery({ workflowId: entityId, filterId });
   const { data: session } = useAuthenticatedUserQuery();
   const caseState = useCaseState(session?.user, workflow);
-  const { mutate: mutateUpdateWorkflowById, isLoading: isLoadingUpdateWorkflowById } =
-    useUpdateWorkflowByIdMutation({
-      workflowId: workflow?.id,
-    });
-  const onMutateUpdateWorkflowById =
-    (
-      payload:
-        | {
-            id: string;
-            approvalStatus: 'approved';
-          }
-        | {
-            id: string;
-            approvalStatus: 'revision' | 'rejected';
-            reason: string;
-          },
-    ) =>
-    () => {
-      if (!payload?.id) {
-        toast.error('Invalid task id');
-
-        return;
-      }
-
-      const action = (
-        {
-          approved: 'approve_document',
-          rejected: 'reject_document',
-          revision: 'ask_revision_document',
-        } as const
-      )[payload.approvalStatus];
-
-      const context = {
-        documents: workflow?.context?.documents?.map(document => {
-          if (document?.id !== payload?.id) return document;
-
-          switch (payload?.approvalStatus) {
-            case 'approved':
-              return {
-                ...document,
-                decision: {
-                  revisionReason: null,
-                  rejectionReason: null,
-                  status: payload?.approvalStatus,
-                },
-              };
-            case 'rejected':
-              return {
-                ...document,
-                decision: {
-                  revisionReason: null,
-                  // Change when rejection reason is implemented.
-                  rejectionReason: payload?.reason,
-                  status: payload?.approvalStatus,
-                },
-              };
-            case 'revision':
-              return {
-                ...document,
-                decision: {
-                  revisionReason: payload?.reason,
-                  rejectionReason: null,
-                  status: payload?.approvalStatus,
-                },
-              };
-            default:
-              return document;
-          }
-        }),
-      };
-      return mutateUpdateWorkflowById({
-        context,
-        action,
-      });
-    };
+  const { mutate: mutateApproveTaskById, isLoading: isLoadingApproveTaskById } =
+    useApproveTaskByIdMutation(workflow?.id);
+  const { mutate: mutateRejectTaskById, isLoading: isLoadingRejectTaskById } =
+    useRejectTaskByIdMutation(workflow?.id);
+  const { mutate: mutateRevisionTaskById, isLoading: isLoadingRevisionTaskById } =
+    useRevisionTaskByIdMutation(workflow?.id);
   const revisionReasons =
     workflow?.workflowDefinition?.contextSchema?.schema?.properties?.documents?.items?.properties?.decision?.properties?.revisionReason?.anyOf?.find(
       ({ enum: enum_ }) => !!enum_,
@@ -103,7 +36,7 @@ export const useCallToActionLogic = () => {
     },
     {
       label: 'Block',
-      value: 'rejected',
+      value: 'reject',
     },
   ] as const;
   const [action, setAction] = useState<(typeof actions)[number]['value']>(actions[0].value);
@@ -114,10 +47,56 @@ export const useCallToActionLogic = () => {
   const onReasonChange = useCallback((value: string) => setReason(value), [setReason]);
   const onActionChange = useCallback((value: typeof action) => setAction(value), [setAction]);
   const onCommentChange = useCallback((value: string) => setComment(value), [setComment]);
+  const isLoadingTaskDecisionById =
+    isLoadingApproveTaskById || isLoadingRejectTaskById || isLoadingRevisionTaskById;
+  const onMutateTaskDecisionById = useCallback(
+    (
+        payload:
+          | {
+              id: string;
+              decision: 'approve';
+            }
+          | {
+              id: string;
+              decision: 'reject' | 'revision';
+              reason?: string;
+            },
+      ) =>
+      () => {
+        if (!payload?.id) {
+          toast.error('Invalid task id');
+
+          return;
+        }
+
+        if (payload?.decision === 'approve') {
+          return mutateApproveTaskById({
+            documentId: payload?.id,
+          });
+        }
+
+        if (payload?.decision === 'reject') {
+          return mutateRejectTaskById({
+            documentId: payload?.id,
+            reason: payload?.reason,
+          });
+        }
+
+        if (payload?.decision === 'revision') {
+          return mutateRevisionTaskById({
+            documentId: payload?.id,
+            reason: payload?.reason,
+          });
+        }
+
+        toast.error('Invalid decision');
+      },
+    [mutateApproveTaskById, mutateRejectTaskById, mutateRevisionTaskById],
+  );
 
   return {
-    onMutateUpdateWorkflowById,
-    isLoadingUpdateWorkflowById,
+    onMutateTaskDecisionById,
+    isLoadingTaskDecisionById,
     caseState,
     action,
     actions,