+++ apps/backoffice-v2/src/pages/Entity/components/Case/hooks/useActions/useActions.tsx
@@ -1,5 +1,5 @@
 import { useCallback } from 'react';
-import { useApproveEntityMutation } from '../../../../../../domains/entities/hooks/mutations/useApproveEntityMutation/useApproveEntityMutation';
+import { useApproveCaseMutation } from '../../../../../../domains/entities/hooks/mutations/useApproveCaseMutation/useApproveCaseMutation';
 import { useDebounce } from '../../../../../../common/hooks/useDebounce/useDebounce';
 import { createInitials } from '../../../../../../common/utils/create-initials/create-initials';
 import { IUseActions } from './interfaces';
@@ -11,23 +11,19 @@ import { useRejectEntityMutation } from '../../../../../../domains/entities/hook
 import { useSelectNextEntity } from '../../../../../../domains/entities/hooks/useSelectNextEntity/useSelectNextEntity';
 import { useWorkflowQuery } from '../../../../../../domains/workflows/hooks/queries/useWorkflowQuery/useWorkflowQuery';
 import { useFilterId } from '../../../../../../common/hooks/useFilterId/useFilterId';
-import {
-  everyDocumentDecisionStatus,
-  safeEvery,
-  someDocumentDecisionStatus,
-} from '@ballerine/common';
 import { useRevisionCaseMutation } from '../../../../../../domains/workflows/hooks/mutations/useRevisionCaseMutation/useRevisionCaseMutation';
-import { Action } from '../../../../../../common/enums';
+import { useCaseDecision } from '../useCaseDecision/useCaseDecision';
 
 export const useActions = ({ workflowId, fullName }: IUseActions) => {
   const onSelectNextEntity = useSelectNextEntity();
   const filterId = useFilterId();
   const { data: workflow, isLoading: isLoadingCase } = useWorkflowQuery({ workflowId, filterId });
-  const { mutate: mutateApproveEntity, isLoading: isLoadingApproveEntity } =
-    useApproveEntityMutation({
+  const { mutate: mutateApproveEntity, isLoading: isLoadingApproveEntity } = useApproveCaseMutation(
+    {
       workflowId: workflowId,
       onSelectNextEntity,
-    });
+    },
+  );
   const { mutate: mutateRevisionCase, isLoading: isLoadingRevisionCase } = useRevisionCaseMutation({
     workflowId: workflowId,
     onSelectNextEntity,
@@ -50,26 +46,7 @@ export const useActions = ({ workflowId, fullName }: IUseActions) => {
   const caseState = useCaseState(authenticatedUser, workflow);
   const { data: users } = useUsersQuery();
   const assignees = users?.filter(assignee => assignee?.id !== authenticatedUser?.id);
-  const hasDecision = safeEvery(
-    workflow?.context?.documents,
-    document => !!document?.decision?.status,
-  );
-  const canTakeAction = caseState.actionButtonsEnabled && hasDecision;
-  // Disable the reject/approve buttons if the end user is not ready to be rejected/approved.
-  // Based on `workflowDefinition` - ['APPROVE', 'REJECT', 'RECOLLECT'].
-  const canReject =
-    canTakeAction &&
-    workflow?.nextEvents?.includes(Action.REJECT.toLowerCase()) &&
-    someDocumentDecisionStatus(workflow?.context?.documents, 'rejected');
-  const canRevision =
-    canTakeAction &&
-    !canReject &&
-    workflow?.nextEvents?.includes(Action.REVISION.toLowerCase()) &&
-    someDocumentDecisionStatus(workflow?.context?.documents, 'revision');
-  const canApprove =
-    canTakeAction &&
-    workflow?.nextEvents?.includes(Action.APPROVE.toLowerCase()) &&
-    everyDocumentDecisionStatus(workflow?.context?.documents, 'approved');
+  const { hasDecision, canApprove, canReject, canRevision } = useCaseDecision();
 
   // Only display the button spinners if the request is longer than 300ms
   const debouncedIsLoadingRejectEntity = useDebounce(isLoadingRejectEntity, 300);