+++ examples/kyb-app/src/domains/workflows/flow-data.type.ts
@@ -1,21 +1,25 @@
 import { ViewsData } from '@app/common/providers/ViewStateProvider/hooks/useViewsDataRepository/types';
+import { BankInformationContext } from '@app/pages/CollectionFlow/components/organisms/KYBView/views/BankInformationView/types';
+import { CompanyActivityConext } from '@app/pages/CollectionFlow/components/organisms/KYBView/views/CompanyActivityView/types';
+import { CompanyInformationContext } from '@app/pages/CollectionFlow/components/organisms/KYBView/views/CompanyInformationView/types';
+import { HeadquartersContext } from '@app/pages/CollectionFlow/components/organisms/KYBView/views/HeadquartersView/types';
 
 export interface PersonalInformationContext {
-  firstName: string;
-  lastName: string;
-  email: string;
-  phone: string;
-  companyName: string;
+  name: {
+    firstName: string;
+    lastName: string;
+  };
+  title: string;
+  birthDate: string;
+  phoneNumber: string;
+  companyCheck: boolean;
 }
 
 export interface DocumentsContext {
   registrationCertificate: string;
   addressProof: string;
-}
-
-export interface BusinessInformationContext {
-  registrationNumber: string;
-  website: string;
+  bankStatement: string;
+  companyStructure: string;
 }
 
 export interface BusinessAddressContext {
@@ -24,21 +28,31 @@ export interface BusinessAddressContext {
 }
 
 export interface UBOSContext {
-  firstName: string;
-  lastName: string;
-  email: string;
+  check: boolean;
+  shareholders: {
+    name: {
+      firstName: string;
+      lastName: string;
+    };
+    title: string;
+    birthDate: string;
+    email: string;
+  }[];
 }
 
 export interface WorkflowFlowData extends ViewsData {
   shared: {
     endUserId?: string;
     businessId?: string;
+    workflowId?: string;
   };
   flowData: {
-    personalInformation: PersonalInformationContext | null;
-    businessInformation: BusinessInformationContext | null;
-    businessAddress: BusinessAddressContext | null;
-    documents: DocumentsContext | null;
-    ubos: UBOSContext[] | null;
+    personalInformation: PersonalInformationContext;
+    companyInformation: CompanyInformationContext;
+    headquarters: HeadquartersContext;
+    companyActivity: CompanyActivityConext;
+    bankInformation: BankInformationContext;
+    companyDocuments: DocumentsContext;
+    ubos: UBOSContext;
   };
 }