+++ packages/ui/src/components/molecules/inputs/DatePickerInput/DatePickerInput.stories.tsx
@@ -1,0 +1,24 @@
+import { useState } from 'react';
+import { DatePickerInput } from './DatePickerInput';
+
+export default {
+  component: DatePickerInput,
+};
+
+export const Default = {
+  render: DatePickerInput,
+};
+
+export const Disabled = {
+  render: () => <DatePickerInput disabled onChange={() => {}} />,
+};
+
+const ControlledComponent = () => {
+  const [timestamp, setTimestamp] = useState(+new Date());
+
+  return <DatePickerInput value={timestamp} onChange={event => setTimestamp(event.target.value)} />;
+};
+
+export const Controlled = {
+  render: ControlledComponent,
+};